+++ src/InvalidJobException.php
@@ -33,7 +33,7 @@ class InvalidJobException extends Exception
         private readonly string $serialized,
         string                  $message = '',
         int                     $code = 0,
-        Throwable               $previous = null
+        ?Throwable              $previous = null
     )
     {
         parent::__construct($message, $code, $previous);