+++ src/drivers/sqs/Command.php
@@ -24,7 +24,7 @@ class Command extends CliCommand
 {
     /**
      * @var Queue
-     * @psalm-suppress PropertyNotSetInConstructor, NonInvariantDocblockPropertyType
+     * @psalm-suppress NonInvariantDocblockPropertyType
      */
     public CliQueue $queue;